LoadThread: Plug file descriptor leaks
The tracefile is opened in TraceParser::open() needs to be closed.
I think here is a reasonable place to do it, when we discover that
we are EOF. The file is still mapped to memory for the purpose
of backtrace displaying but the mmap region should not be affected
by the fact that the fd used is closed.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>