StringPool: Improve the cutoff policy
Earlier we considered a hash value to be a lost cause if we were
merely over the cutoff limit. Here we add the additional condition
that we must have had more new allocations than sucessful lookups
in the AVL-tree, in order to consider it a lost cause.

I expect the impact of this to be little, it could have some
impact on corner case traces.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>