MainWindow, TraceAnalyzer: Find wakeup in filtered events
This fixes a bug that when the events view is filtered, we will not
correctly find the wakeup event there, even if it is among the
filtered events. This will attempt to do it and scroll the events
view to the wakeup event if it is included, if it is not, there will
be no scrolling.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>