New Feature: event type filtering of the event view
Previously I have implmented the feature that the event view cna be
filtered by task, or pid, actually tid. Now with this it becomes
possible to filter by event type.

The button uses the same icon as the task selector, there is a need to
draw a new icon but I have no time for it just now.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>