Improve filtering
This modification contains to major improvements of the filters.

1. It adds the posibility of doing OR based filtering. That means
that when you select "||" instead of "&&", then the result would
be the union of any other filtering, rather than the intersection.

2. It adds a checkbox to task based filtering where one can chose
that the task based filtering includes wakeups, forks and
sched_switch events where the filtered tasks are targeted, even
if the event itself belongs to some other task.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>