AVLTree: Make the compare function more flexible
This makes the AVLTrees behave in the following manner:

If no compare class is given as third argument to the template,
then they keys are compared with the '==' and '<' operators.

If there is a compare class without a compare() function, then
the keys will also be compared as above.

If there is a compare class with the compare() function ,then
that function will be used.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>