Improve the timefilter icon
Logically, the cursor lines should be solid, I guess because of the
solid lines in the move cursor icons. However, since the cursors
are really dashed lines and the fact that I from the very first time
I saw this icon on the street has had the idea that they ought to be
dashed lines, in order to look more sophisticated and to get that
correct 1980s feeling.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>