MainWindow: Add a new feature to export events
This is a new feature that makes it possible to export filtered
events. In my thinking, the most important use case for this is
to be able to generate flame graphs of a certain subsets of
events, for example events that are limited to a certain
time interval, event type (probably cycles), and task.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>