TraceAnalyzer: Deal with negative pid in perf events
In the 4.14.x kernel series, events of this type has seen the light
of the day:

:-1    -1 [002]  3165.146420:                     sched:sched_switch:

Here the pid is negative, we need to deal with it somehow. We will
still assume that no "normal task" will have negative pid. These
events only occur after a task has exited, it's the switch event after
a task has exited.

The policy will be that we will not take negative pids seriously,
except as pids of events. We will assume that no task that the
user is really interested in will ever have a negative pid.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>