Use the vtl::Time class for time
This will greatly reduce the usage of floating point values (doubles)
to store time values. The vtl::Time is a fixed point time
implementation.

However, doubles are still used for the displaying of graphs, since
QCustomPlot expects coordinates in double.

The purpose of this change is not to improve performance but to
make it easier to display time consistently. Specifically to
ensure that time is displayed with the same precision as found in
the trace and rounded in an expected way, e.g. that when the user
clicks on a time value, the exact same value is seen in the
CursorInfo field.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>