Add support to compile with clang
This change adds a commented out option to traceshark.pro that will
cause the program to be compiled with clang. Uncomment this to use
clang:

Also, we add build information to the mainwindow, under
"Help" -> "About Traceshark". It should compile the compile that was
used as well as the build date and time.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>