LegendGraph: Remove dead code
These LegendGraph objects are no longer required. They were part of
solution for legends with the earlier version of QCustomPlot but
now this is accomplished with the TaskGraph::fromQCPGraph()
function. This function can be used to convert a QCPGraph pointer
to a TaskGraph pointer, so that we can now to what task a certain
QCPGraph belongs.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>