Setting: Change name of MAX_SETTINGS
There is an off by one bug in the checking of the bounds in
Setting:clicked(). I believe that a reason for this mistake was the
misleading name of the MAX_SETTINGS, so we change it to NR_SETTINGS.

Nota bene, I am not aware of any practical use cases in which this bug
could be triggered by the end user.

This bug was found by Coverity. It is CID 1469883.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>