Fix problem with exporting filtered events
It seems like it's unwise to use O_DIRECT flag for ext4 filesystems.
At least it doesn't work with 4.17.3

This caused the functionaly of exporting a filtered events to not work
if the user was on an ext4 filesystem.

I didn't notice this problem because when I implemented this I was
a system with btrfs.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>