vtl::Time: Change the internal representation to single 64-bit
It should be possible to get rid of having two unsigned ints, that
in practice will be 32-bit. Instead we hope that we are using a
compiler that supports 64-bit ints. Most 32-bit platforms have
a compiler with support for 64-bit ints. This will make the math
functions much simpler to implement.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>