MainWindow: Limited the allowed areas for QDockWidgets
I feel that the application will become too complex if we allow the
user to move around these dialogs, actually they are widgets now, not
dialogs. The complexity arrises when we want to do automatic tabifying
of those dialogs that by default go to the same docking area, i.e.
the task selector and the global statistics (taskSelectDialog, and
statsDialog.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>