Revert "AbstractTask: Use unsigned int for indices"
I changed my mind again. Let's use signed arithmetic for all indices.
The justification is that underflow detection is easier, e.g.

for (index = start; index >= 0; index--) {
}

Furthermore, we would with the present implementation not be able to
benefit from having one extra bit. The eventswidget would stop working
when the index (size) beomes larger than INT_MAX because the
QAbstractTableModel, which is indirectly used through EventsModel,
uses signed int for its rows, which corresponds to the index in the
events TList.

What this means is that Traceshark will on most platforms only be able
to handle traces that contains a maximum of 0x7fffffff, or 2147483647,
events. If we assume that each events is about 1000 bytes then it
still means that we could handle traces that are about 2 Petabytes,
assuming that we are not hit by any other limitation such as the
amount of RAM in the computer.

This reverts commit c4ff9ff8b845eec173c2ab92952b7678b40b2b63.