parser/perf: Improve compatibility with distro perf
This is an attempt at improving the out of the box compatibility
various Linux distros, specifically Debian and Fedora. The idea
is to make traceshark compatible with the perf version that is
shipped with these distros.

One of the reasons for this is that it seems that for newer releases,
the distro perf works better than the perf that comes with the mainline
kernel.

This modification make the perf param functions understand the
following format of sched_swtich:

LoadThread  5182 [000]  5546.159804: \
sched:sched_switch: LoadThread:5182 [125] D ==> parserThread:5184 [125]

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>