TraceAnalyzer: Improve handling of missing scheduling event
If we have a missing scheduling event, we may face the situation that
we are missing the event when a task should have been scheduled out.

The handling of time was buggy in the previous implementation:
prevtime = cpuTask->schedTimev.last();

Here the left side is vtl::Time and the right side is double.
Probably the vtl::Time() class should be improved so that this
would either work as expected or generate a compile time
failure.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>