FtraceGrammar: Make the parseLine() non-destructive
In the beginning of the trace, which may last for a long time, both
Ftrace and perf parser may consume the same TraceLine objects,
therefore it is necessary that they are non-destructive, in order
to not interfere with each other.

This fixes a very annoying bug when viewing perf traces from the
4.14.x kernel series (other version may be affected as well).

The impact is that the final switch of zombie functions are lost:
:-1    -1 [008] 15589.436740:                \
sched:sched_switch: prev_comm=QFileInfoGather prev_pid=7323 \
prev_prio=125 prev_state=Z ==> next_comm=swapper/8 next_pid=0 \
next_prio=120

The negative pid causes the matching of the Ftrace parsing to go so
far that the strings contained in the TraceLine objects are modified,
which in turn causes the perf parsing to fail.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>