MainWindow, TaskInfo, InfoWidget: Reschuffle UI
We will introduce a new TaskToolbar, so that the UI can work also for
screens that doesn't have a large horizontal resolution.

The behavior will be dependen on horizontal screen resolution; we will
use more vertical space for the task toolbar in the case that our
screen is not wide enough.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>