MainWindow: Only enable the wakeup button if a task is selected
The wakeup button is used to search for a wakeup of the currently
selected task. For this reason, it makes sense to only enable it
if we have selected a task. This is not a huge change in
functionality, previously it was possible to push the button but
nothing happened. Now the button will be disabled. The idea here
is merely to give the UI a better feel.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>