MainWindow: Deselect task if the wakeup event has pid 0
If the wakeup event is run with pid 0, i.e. swapper, then we will
deselect the currently selected task and remove it from the task
toolbar.

The reason for this change is to make the UI more logical and
consistent. The idea is that after the button is pushed, we
should have automagically selected the task that that caused the
wakeup. If the wakeup event can't be attributed to a task, then
nothing should be selected.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>