New feature: Add a button to find a sched_waking event
This adds a new button that can be used to find the waking event
that precedes a wakeup event. The button is only enabled when a
a wakeup event is selected in the events view. The wakeup event
can be selected by scrolling the events view and selecting an
event, or more typically, one can use the find wakeup event button
for that.

This feature is useful when a task is woken up from another CPU.
In that case, the backtrace of the wakeup event will only point to
something like this:

2863f8 ttwu_do_wakeup
286522 ttwu_do_activate
2878a7 sched_ttwu_pending
29fec6 do_idle
2a00ff cpu_startup_entry
237418 start_secondary
2000d5 [unknown]

Then, when a backtrace like this is seen, one should push the new
button, the icon is a green alarm clock, then the "sched_waking"
event will be found and it will hopefully have a backtrace that
points to the real wakeup source.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>