parser: Deploy the NullStr
Many parsing functions might return nullptr, in case the argument
format of an event has an unexpected format.

These nullptr results would then be used by the analyzer code, whch
would crash the program. I expect that checking all the returns
for nullptr would have an impact on performance.

Instead, we introduce the NullStr, this would then appear as
"parsing error" strings as task names in the UI, instead of
crashing the program.

If we want, we can later add more elaborate error checking code
by having the analyzer code checking for NullStr return values,
and showing an errror dialog to the user.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>