New feature: find previou waking event of the selected task
This adds a new button that combines the functionality of the already
existing "Find Wakeup" and "Find Waking" buttons.

The "Find Wakeup" button finds the preceding wakeup event of the
currently selected task. The "Find Waking" button finds the waking
event of the currently selected wakeup event. It is common then
to first use the "Find Wakeup" button and then the "Find Waking"
button. This new button combines that functionality into one new
button, so that we can directly find a preceding waking event of the
currently selected task, without first findind the wakeup event.

However, it assumes that there are a waking event. That is not
universally true, because in some cases there is a wakeup without
a waking. One known example is when a new task is started. The
sched_wakeup_new events have no corresponding sched_waking events.
I have not verified whether there are other cases.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>