New feature: Add a button to filter on the currently selected task
This is an obvious missing feature. It's very handy to be able to
filter on the currently selected task. The filtering will be
inclusive, that is, it will also match sched_wakeup, sched_wake_new,
sched_process_fork and sched_switch events that involves the selected
task, even if these events are executed with another pid.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>