MainWindow: Only enabe task actions if a task is selected
This is a modification of behavior so that those actions, both menu
items and related buttons, only are enabled if a task is selected.

This affects the actions to:
- add a task to the legend
- add a task graph of a task
- remove a task graph of a task
- filter on a task

I am not 100% sure of this change. Might require a bit of testing
to see if it is possible to confuse the detection of whether a task
is selected or not. Anyway, the find wakeup action has for a long
time been using this.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>