New feature: find sched_switch sleep event
This new button will find the next sched_switch event that puts the
currently selected task to sleep. It's sometimes useful to be able
to see what kind of event makes a particula task to go to sleep.

A typical way to use this would be to select a task, push the button,
and then double click on the 'Info' field of the sched_switch event,
in order to get the backtrace.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>