README.md: Always use image from master branch
After considering this, I have decided that it's better if the images
are always taken from the master branch, so that if an image is
updated, we will automagically get the latest image.

The downside is of course that if someone renders a historical
README.md, he will not see the historical images but the latest ones
from the master branch.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>