New feature export CPU cycles events
This buttons makes it unnecessary to filter on cpu-cycles or cycles,
instead we can just export these events directly with one button.
This convenience button is in recogniction of the fact that exporting
is mostly done to generate a flame graph and it ought to be as quick
and easy as possible.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>