TraceFile: Do not mmap() the trace file permanently
In order to obtain the backtraces, we used the pointer in the
event.postEventInfo->ptr, it points to the mmap() that is created
in TraceFile::TraceFile.

The problem with this approach is that it mmaps the whole trace file
permanently; on 32-bit platforms this means that we will quickly run
out of address space.

In order to fix this, we change the postEventInfo to the new Chunk
type, which consists of offset and length. Thus, we can find
backtraces by using lseek64() and read().

The problem with this new approach is that when exporting a file and
we need to execute two syscalls for every backtrace. For this reason,
TraceFile has the two functions, allocMmap() and freeMmap(). When
exporting we use, allocMmap() to try to allocate a mmap() for the
tracefile, when done, we free it with freeMmap(). This may fail on
32-bit platforms, but in that case lseek64() and read() will be used
as fallback. Thus, we will be able to use mmap() when it succeeds,
for small files on 32-bit platforms and for all practical files on
64-bit platforms.

The big improvement with this is that I can open a 1.6 gigabyte
file with traceshark on Raspberry PI 3, with memory to spare.
Previously this failed when doing the mmap() in
TraceFile::TraceFile() not because a lack of memory but because the
address space was exhausted.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>