MainWindow: Improve task selection implementation
This is an attempt to refactor the task selection code. It concerns
both the selection that happens when any of the find wakeup/waking
buttons are used and the dialogs that can select tasks.

Previously similar code was duplicated three times. This attempts to
use the same selection code in all three cases.

In addition, there is an improvement in the logic. Now we always try
to find a task graph first, if one is visiable. If no task graph is
available, we will use a cpu graph. When finding a wakup or waking
event, the selected cpu task will be on the cpu where the waking or
wakeup action ran. However, in the case of selecting a task through
one of the statistics dialogs, or the task selection dialog, then
the cpu graph with the highest number of sched_switch events.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>