MainWindow::addPreemptedGraph(): Fix empty check
The code checed if the wrong vector is empty. The problem came when
the function was copy pasted from MainWindow::addStillRunningGraph,
and the empty test was not adjusted.

Perhaps these functions should be generalized. That will be a later
problem. For now, we only fix the buggy behavior and leave the
cosmetic problem.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>