TraceFile: Fix problem with uninitialized variable
This mappedFileSize was used but not initialized. We change the name
to fileSize and the type to int64_t and make sure that it is
initialized. At the same time we make the getFileSize() function to
be inlined, since now it's just returning the value of a member
variable.

This is CID 1476429 in Coverity.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>