qcustomplot: Import new revision
Beware, this may be buggy!

!!!!!!!!!!!!!!!!!!!!!!!! ATTENTION !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
You might want to go back to the latest release, if you get problems
with plotting or other weird problems. The latest release is:
v0.9.4-beta.
!!!!!!!!!!!!!!!!!!!!!!!! ATTENTION !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!

Import the newest revision of the rel-2.0.1-traceshark branch at
git@gitlab.com:cunctator/QCustomPlot.git. The revision is
eeb16a9f03800ecc78f4d16fb3618ec0a0923985.

It contains the followin performance optimization:

QCPLayer: Use QCPList instead of QList

The mChildren member for QCPLayer becomes a problem when we have a
very large number of plottables. For example the QList<T>::contains()
operation is linear with respect to the number of elements.

In order to improve the situation, with a fairly small effort, we
introduce this new QCPList<T> class. It is a doubly linked list
but it also contains a QMultiMap<> that can be used to implement
faster contains() and removeOne() operations.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>