MainWindow: Use thicker lines if opengl is enabled
According to my testing, the only benefit from opengl is if we are
drawing the scheduling graphs with thicker lines. With this change
we will draw scheduling graphs that are 2 pixels wide if opengl is
enabled.

Over time, I plan to make it possible to change the line width from
the UI, if opengl is enabled. For now, the idea is that users who
want thicker lines should simply enable opengl.

There is no benefit from opengl with the 1 pixel lines and it's not
possible to enable 2 pixel lines without opengl, the performance
penalty is too great.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>