GraphEnableDialog, MainWindow: Do not display excessive migrations
If a trace contains an excessive number of migrations, then
traceshark performance will suffer. For example, it is going to
take a long time to destroy all those migration arrows when the
trace is closed, because each of them is a QObject.

For this reason, we will by default only display migrations if their
number is less than 200000. There is a new setting in the graph enable
dialog, "Unlimited migrations", that can be used to display the
migrations, regardless of how many they are. However, this setting is
disabled by default.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>