MainWindow: Improve policty for default line width and opengl
We should only enable OpenGL by default, if it has been enabled during
compile time and we are on a high resolution screen. There is no use
to enable OpenGL on a low resolution screen because there we only want
one pixel wide lines and they can better be drawn with software
rendering.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>