Enable OpenGL by default
I know feel as if there are sane policies in place to deal with
OpenGL. We enable it if we are on a high resolution screen and
then we draw the scheduling graphs with a pixel width of 2. This
is useful for most people, since high resolution screens are
becoming quite common and using a width of 1 feels a bit too tiny.
The user can still change the width or disable OpenGL.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>