MainWindow: Fix bug when task is selected by a dialog
The state of the remove and add task graph buttons, whether they
should be enabled or not was not updated when selecting a task
through on of the dialogs. In the task select dialog and in the
statistics dialog a task can be selected by double clicking and
in this case the state was not correctly updated.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>