StringPool: Drop the hash parameter from the interface of searchAllocString()
It's a daft thing to require the user of searchAllocString() to
calculate the hash value. Instead, we turn StringPool into a template.
The template HashFunc parameter is a functor that determines which
hash function to use.

The template parameter HashFunc has a default parameter,
DefaultHashFunc, which simply calls StrHash32, which so far is the
only string has function used in this project.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>