Setting: move global settings to new class SettingStore
Previously, the program had a global state that could be manipulated
through some of the static get* and set* functions in the Setting
class. Now this setting state is moved into a new class that is
named SettingStore. An instance of this new class is created in
MainWindow and pointers are passed to TraceAnalyzer and to
GraphEnableDialog.

This would make it easier in the future to modify the program so that
it would be possible to have several MainWindow instances, since there
would now be less global state to cope with. I have currently no plans
to enable the user to have multiple MainWindow instances in the same
process but I think it's always a good idea to avoid having a lot of
global state.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>