GraphEnableDialog: Introduce the ValueBox
The ValueBox is a new generalized input widget that can aquire either
boolean or integer values from the user, depending on the type of the
value of the setting that it is bound to. The main benefit with this
is that it enabled us to make a much cleaner implementation of the
GraphEnableDialog class.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>