perfparams.h: Increase robustness perf_sched_process_fork_*
This improves the robustness of perf_sched_process_fork_childpid() and
perf_sched_process_fork_parent_pid.

I leave perf_sched_process_fork_childname_strdup as it is, considering
that it is not currently in use by the analyzer.

Of these functions perf_sched_process_fork_childpid() is the only on
that is currently in use by the analyzer.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>