SettingStore: Move the setting for max vertical wakeup latency
It makes sense to move this so that it is displayed after the settings
that it is dependent on.

It would not be necessary to move the functions in the SettingStore()
constructor, only the order in the enum matters. However, the code
looks a bit pretties if they are in the same order.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>