SettingsStore: Change the disabled value of DEFAULT_MAX_VRT_LATENCY
Since we cannot make the spinbox display a value that is out of range,
we just display the default value when this is not enabled, even if
the value is not used. Displaying 0 would in some ways make more sense
but what happens with the current framework is that the value gets
rounded up to 1, which is the minimum value.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>