MainWindow: Add a diagnostic message for slow closeTrace()
This change generates a diagnostic message to the console whenever
closing the trace takes more than one second. This happens
"normally" when we have a longer trace with thousands or tens of
thousands of tasks.

This is intended to ease development in the future, if I should
want to dabble with trying to improve the closing time of large
traces.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>