MainWindow: Implement a vertical scrollbar for the plot
If there are many CPUs in the trace and/or the screen size is small,
then the plot will not be readable in the vertical direction. This
problem is exacerbatet by the fact that the vertical zoom and scroll
must be enabled in an obscure way.

This change makes the default zoom to only display a part of the trace
in the vertical direction and enabling a scrollbar.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>