MainWindow: Add fullzoom and modify the default zoom action
With this change, fullzoom will have the same functionilty and icon
as the old default zoom. The new default zoom will try to avoid
zooming out too much in the vertical direction. If many there are too
many CPUs in the trace, then it will zoom out less in the vertical
direction and the user will need to either use the scrollbar, or
manually zoom out, for example by clicking the full zoom button or
selecting the vertical axis.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>