Add a new action for enabling vertical zoom and scroll
Currently, the only way to enable vertical zoom and scroll is to select
the vertical axiis. This is unfortunately so obscure that many users
will not find it.

This commit adds a new button and item in the view menu, so that the
user can select vertical zooming/scrolling without knowing about the
trick with selecting the vertical axis. Both the button and the menu
item are checkable and the checed state is synced with the selection
of the vertical axis.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>