main(): Add Qt::AA_UseHighDpiPixmaps
I believe that it perhaps is better to add this option to get high dpi
icons. I am not sure whether the icons will look better or not, since
the icon pixmap handling is done by Qt rather than the application
code. Also, I don't have any HiDPI system to test it with.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>