SettingStore: Always let the user set the frequency line width
With this change we will allow the user to select the frequency line
width, even if the frequency graphs are not displayed.

The reason for this change is that I want to make it consistent with
scheduling and idle graphs, which can have their line widths set even
if these graphs are not being displayed.

We cannot [easily] make scheduling and idle line width dependent on
whether the graphs are displayed, because they are already depending
on the OpenGL setting and the setting system as currently implemented
cannot deal with a setting depending on more than one other setting.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>