vtl/time.h: Fix detection of integer limits
The detection of limits of various integer types and thereby the
selection of a long enough integer type for timeint_t was broken with
the commit 75ffc72e3cefb821138f489d0441a4eb24a90d2c
"Standardize usage of underscore"

This resulted in some 32-bit platforms selecting a too short (32-bit)
integer type. This change should fix the bug.

I believe that this is an important bugfix, at least for 32-bit
platforms.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>