Implement the new UI actions for the currently selected events
Let's add a new menu for the currently selected event, the items in
this menu can then be used for operations on the surrently selected
event.

One of these already existed but it's too obscure; many users will
probably not notice that it's possible to display the backtrace by
double clicking on the info field. Because of this we will create
an explicit menu item for it.

Futhermore we will add meny items for the following operations:
- Move the red cursor
- Move the blue cursor
- Filter on the event cpu
- Filter on the event pid
- Filter on the event type

It is also possible to trigger the above actions by double clicking on
the relevant column in an event in the events widget. If the time
column is double clicked, then the active cursor is moved.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>