EventsWidget: Change the selection policy to single row selection
In the idea case we should only let the user select a single cell. There
doesn't seem to be any reason why a user would need to select multiple
cells.

However, I do not know to use the API to select a single cell, this
would be needed when we move the cursor.

For this reason we make the EventsWidget so that it is only possible
to select single rows. In this case we can use the selectRow() in
QTableView.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>