bugfix: Do not enable wide graphs if OpenGL is not available
This is a bug that mostly affects distros that build traceshark to link
against a system libqcustomplot library where OpenGL has been disabled.

The problem is that we are using wide lines, that is line widths that are
greater than 2, when no OpenGL support is available. This causes very bad
scrolling and zooming performance.

The problem could also be solved by disabling OpenGL at build time by
uncommenting the following line in traceshark.pro:
 # DISABLE_OPENGL = yes

However, it seems that if a misconfiguration is possible it will also
happen. Thus, this change fixes the problem at run time.

This change does the following:
- If enabling OpenGL fails, then we recheck the settings that depend on
  OpenGL.
- When we load settings from the .traceshark file, we check that no
  dependencies are broken.
- When we intialize the values in the SettingStore constructor we also
  check that no dependencies are broken.

All in all this change should make it harder if not impossible to end up
in a situation where we draw wide lines without OpenGL, or have any
other type of setting combination that should not be allowed because
of dependencies.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>