Make MainWindow the parent of dialogs
I believe that these dialogs should have MainWindow as their parent. This
solves the problem with it being possible to close the MainWindow while
still having a dialog open with no way of opening a new MainWindow. This
is of course a very minor problem but could be confusing for some users.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>