MainWindow: Try to display the right copyright years
This is an attempt at displaying the correct copyright years for
QCustomPlot. We assume that it's v2.0.0, v2.0.1, or v2.1.0. It's likely
difficult to compile traceshark with an older version QCustomPlot and
we cannot possibly know about any future versions. Therefore, if the
version is not 2.0.1, or 2.0.0, we display the years for 2.1.0, so that
for uknown version we display the years for 2.1.0.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>