TraceAnalyzer: Use the time of the sched_switch event for wakeup latency
This changes the behavior so that when the wakeup latency is calculated,
it is what most people would expect, the time between the wakeup event
and the sched_switch event.

Previously, we used the time of the sched_switch event + FAKE_DELTA.
FAKE_DELTA is just an artistic estimate of the time it takes to switch
between tasks. It's used for rendering purposes so that there will be
a small gap between a new task and the old task.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>