TraceAnalyzer: Implement recording of scheduling latencies
This change will create an additional member of the TraceAnalyzer class,
vtl::TList<Latency> wakeupLatencies;

This member is just a TList that contains the scheduling latencies of the
trace. Initially it will be sorted by magnitude of the latency. The Latency
class has a place member, which tells the placement of the latency in the
TList. This aids us in quickly resorting the TList, depending on user
preference.

There is not yet any code for displaying the computed latencies, so
currently this is not really visible to the user. It will be future work to
implement the UI functionality for displaying this.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>