analyzer: Improve naming for scheduling latency code
These "wakeup" things should really be called scheduling delay, or latency,
because they do not only include the wakeup latencies but also the latency
for a long running task that has been scheduled out and was still runnable
when it was scheduled out.

I currently plan to add real wakeup latency code soonish, and for this
reason we need to rename these, in order to avoid confusing naming in the
future.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>