analyzer: add recording of wakeup latencies
This is meant to be real wakeup latencies, that is only the latencies
between the waking up of a task and it being scheduled. If a task is
still runnable when scheduled out, that is not counted.

There is not yet any code to display these latencies. That will be
future work.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>