New feature: Add a new widget for displaying latencies
This is a new feature that adds a copule of new icons and menu entries for
displaying wake and scheduling latencies.

The wakeup latencies are strictly the time between wakeup and a task being
schedule, while the scheduling latencies also include the case wher a task
has been scheduled out in a runnable state.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>