SettingStore: prevent deprecated settings from being shown by UI
The deprecation of the old settings is buggy. They are currently being
shown to the user and without proper dependencies or descriptons, although
they were meant to exist only for extraction from .traceshark.

For this reason let's change NR_SETTINGS to be smaller than the
deprecated settings and introduce the new NR_ALL_SETTINGS for the
case when we want to refer to all settings, not just those that
are to be displayed.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>