time.h: Clean up the constructor and the min/max macros
The VTL_TIME_MAX and VTL_TIME_MIN macros are  buggy, let's fix them.
Luckily, they are not used anywhere, so this is more of a cosmetic bug
fix and to make them usable in the future.

The Time() constructor is also changed to be less chaotic. Having a
separate argument for seconds is not that great when we are close to
the maximum or minimum values, if for no other reason that it creates
an impression for the user of the class that the internal representation
would be for both seconds and nanoseconds, i.e. that an overflow would
be extremely unlikely, while in reality the internal representation is
nanoseconds only and an overflow is more possible.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>