MainWindow::exportLatencies(): Allow the user to override the format
The filter parameter of the  QFileDialog::getSaveFileName() was used in a
suboptimal way.

This change modifies the behavior so that the user can use the file saving
dialog to change his mind about the format in which to save the latencies.

It would be possible to only use this file saving dialog to select the
format but I am afraid that many users would miss the fact that there are
several options available there in the file saving dialog's filter combo
box.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>