TraceAnalyzer: Handle sched_switch events with identical timestamps
It can happen that consecutive sched_switch events have identical timestamps.
Theoretically, it can also happen that the timestamps would be closer than
2 * FAKE_DELTA, which would make the FAKE_DELTA hack work badly.

This change add another hack that handles the case with consecutive identical
timestamps, or timestamps that are too closer. If it happens, we simply add
time to any consecutive sched_switch events with the same or too close
timestamp.

Also change FAKE_DELTA from 20ns to 50ns and add a comment to explain what
FAKE_DELTA is for.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>