analyzer: make sure that the lastRunnable is valid
In some cases, if we lose a sched_wakup or sched_wakeup_new event, it is
possible that we will use a nonsense value for lastRunnable, resulting in
ridiculously long latency values in the graphs.

In particular, if we lost the sched_wakeup_new event that comes directly
after a sched_process_fork event, then we will get a very long wakeup graph
if we add a unified graph.

We solve this by replacing the boolean variable lastRunnable_is_sched by
a enum lastRunnable_status. This can either be invalid, wakeup, or sched.
The idea is that we track if the current task:
- is not runnable, that is runnable is invalid.
- has been woken up by a wakeup event that made it runnable
- has been shceduled out in a runnable state.

We need to distinguish between wakeup and scheduled out in a runnable state
because we are calculating different kind of statistics, both scheduling
latencies and wakeup latencies.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>